Продвинутый FrustumCulling

+++ src/world/World.h
@@ -15,9 +15,9 @@ class World {
 public:
 	std::string name;
 	WorldFiles* wfile;
-	int seed;
+	uint64_t seed;
 
-	World(std::string name, std::filesystem::path directory, int seed, EngineSettings& settings);
+	World(std::string name, std::filesystem::path directory, uint64_t seed, EngineSettings& settings);
 	~World();
 
     void write(Level* level, bool writeChunks);
